refactor: delete object.js, change lf
