+++ src/index.html
@@ -8,6 +8,6 @@
     <title>Virtual keyboard</title>
 </head>
 <body>
-<script type="module" src="script.js"></script>
+<script src="script.js"></script>
 </body>
 </html>
