+++ src/script.js
@@ -1,4 +1,274 @@
-import keys from './object.js';
+const keys = {
+  enKeyLowerCaseFirstRow: {
+    Digit1: '1',
+    Digit2: '2',
+    Digit3: '3',
+    Digit4: '4',
+    Digit5: '5',
+    Digit6: '6',
+    Digit7: '7',
+    Digit8: '8',
+    Digit9: '9',
+    Digit0: '0',
+    Minus: '-',
+    Equal: '=',
+  },
+  enKeyLowerCaseSecondRow: {
+    KeyQ: 'q',
+    KeyW: 'w',
+    KeyE: 'e',
+    KeyR: 'r',
+    KeyT: 't',
+    KeyY: 'y',
+    KeyU: 'u',
+    KeyI: 'i',
+    KeyO: 'o',
+    KeyP: 'p',
+    BracketLeft: '[',
+    BracketRight: ']',
+    Backslash: '\\',
+  },
+  enKeyLowerCaseThirdRow: {
+    KeyA: 'a',
+    KeyS: 's',
+    KeyD: 'd',
+    KeyF: 'f',
+    KeyG: 'g',
+    KeyH: 'h',
+    KeyJ: 'j',
+    KeyK: 'k',
+    KeyL: 'l',
+    Semicolon: ';',
+    Quote: "'",
+  },
+  enKeyLowerCaseFourthRow: {
+    KeyZ: 'z',
+    KeyX: 'x',
+    KeyC: 'c',
+    KeyV: 'v',
+    KeyB: 'b',
+    KeyN: 'n',
+    KeyM: 'm',
+    Comma: ',',
+    Period: '.',
+    Slash: '/',
+  },
+  enKeyUpperCaseFirstRow: {
+    Digit1: '!',
+    Digit2: '@',
+    Digit3: '#',
+    Digit4: '$',
+    Digit5: '%',
+    Digit6: '^',
+    Digit7: '&',
+    Digit8: '*',
+    Digit9: '(',
+    Digit0: ')',
+    Minus: '_',
+    Equal: '+',
+  },
+  enKeyUpperCaseSecondRow: {
+    KeyQ: 'Q',
+    KeyW: 'W',
+    KeyE: 'E',
+    KeyR: 'R',
+    KeyT: 'T',
+    KeyY: 'Y',
+    KeyU: 'U',
+    KeyI: 'I',
+    KeyO: 'O',
+    KeyP: 'P',
+    BracketLeft: '{',
+    BracketRight: '}',
+    Backslash: '|',
+  },
+  enKeyUpperCaseThirdRow: {
+    KeyA: 'A',
+    KeyS: 'S',
+    KeyD: 'D',
+    KeyF: 'F',
+    KeyG: 'G',
+    KeyH: 'H',
+    KeyJ: 'J',
+    KeyK: 'K',
+    KeyL: 'L',
+    Semicolon: ':',
+    Quote: '""',
+  },
+  enKeyUpperCaseFourthRow: {
+    KeyZ: 'Z',
+    KeyX: 'X',
+    KeyC: 'C',
+    KeyV: 'V',
+    KeyB: 'B',
+    KeyN: 'N',
+    KeyM: 'M',
+    Comma: '<',
+    Period: '>',
+    Slash: '?',
+  },
+  enForCapsSecondRow: {
+    KeyQ: 'Q',
+    KeyW: 'W',
+    KeyE: 'E',
+    KeyR: 'R',
+    KeyT: 'T',
+    KeyY: 'Y',
+    KeyU: 'U',
+    KeyI: 'I',
+    KeyO: 'O',
+    KeyP: 'P',
+    BracketLeft: '[',
+    BracketRight: ']',
+    Backslash: '\\',
+  },
+  enForCapsThirdRow: {
+    KeyA: 'A',
+    KeyS: 'S',
+    KeyD: 'D',
+    KeyF: 'F',
+    KeyG: 'G',
+    KeyH: 'H',
+    KeyJ: 'J',
+    KeyK: 'K',
+    KeyL: 'L',
+    Semicolon: ';',
+    Quote: '\'',
+  },
+  enForCapsFourthRow: {
+    KeyZ: 'Z',
+    KeyX: 'X',
+    KeyC: 'C',
+    KeyV: 'V',
+    KeyB: 'B',
+    KeyN: 'N',
+    KeyM: 'M',
+    Comma: ',',
+    Period: '.',
+    Slash: '/',
+  },
+  ruKeyLowerCaseSecondRow: {
+    KeyQ: 'й',
+    KeyW: 'ц',
+    KeyE: 'у',
+    KeyR: 'к',
+    KeyT: 'е',
+    KeyY: 'н',
+    KeyU: 'г',
+    KeyI: 'ш',
+    KeyO: 'щ',
+    KeyP: 'з',
+    BracketLeft: 'х',
+    BracketRight: 'ъ',
+    Backslash: '\\',
+  },
+  ruKeyLowerCaseThirdRow: {
+    KeyA: 'ф',
+    KeyS: 'ы',
+    KeyD: 'в',
+    KeyF: 'а',
+    KeyG: 'п',
+    KeyH: 'р',
+    KeyJ: 'о',
+    KeyK: 'л',
+    KeyL: 'д',
+    Semicolon: 'ж',
+    Quote: 'э',
+  },
+  ruKeyLowerCaseFourthRow: {
+    KeyZ: 'я',
+    KeyX: 'ч',
+    KeyC: 'с',
+    KeyV: 'м',
+    KeyB: 'и',
+    KeyN: 'т',
+    KeyM: 'ь',
+    Comma: 'б',
+    Period: 'ю',
+    Slash: '.',
+  },
+  ruKeyUpperCaseFirstRow: {
+    Digit1: '!',
+    Digit2: '"',
+    Digit3: '№',
+    Digit4: ';',
+    Digit5: '%',
+    Digit6: ':',
+    Digit7: '?',
+    Digit8: '*',
+    Digit9: '(',
+    Digit0: ')',
+    Minus: '_',
+    Equal: '+',
+  },
+  ruKeyUpperCaseSecondRow: {
+    KeyQ: 'Й',
+    KeyW: 'Ц',
+    KeyE: 'У',
+    KeyR: 'К',
+    KeyT: 'Е',
+    KeyY: 'Н',
+    KeyU: 'Г',
+    KeyI: 'Ш',
+    KeyO: 'Щ',
+    KeyP: 'З',
+    BracketLeft: 'Х',
+    BracketRight: 'Ъ',
+    Backslash: '/',
+  },
+  ruKeyUpperCaseThirdRow: {
+    KeyA: 'Ф',
+    KeyS: 'Ы',
+    KeyD: 'В',
+    KeyF: 'А',
+    KeyG: 'П',
+    KeyH: 'Р',
+    KeyJ: 'О',
+    KeyK: 'Л',
+    KeyL: 'Д',
+    Semicolon: 'Ж',
+    Quote: 'Э',
+  },
+  ruKeyUpperCaseFourthRow: {
+    KeyZ: 'Я',
+    KeyX: 'Ч',
+    KeyC: 'С',
+    KeyV: 'М',
+    KeyB: 'И',
+    KeyN: 'Т',
+    KeyM: 'Ь',
+    Comma: 'Б',
+    Period: 'Ю',
+    Slash: ',',
+  },
+  ruForCapsSecondRow: {
+    KeyQ: 'Й',
+    KeyW: 'Ц',
+    KeyE: 'У',
+    KeyR: 'К',
+    KeyT: 'Е',
+    KeyY: 'Н',
+    KeyU: 'Г',
+    KeyI: 'Ш',
+    KeyO: 'Щ',
+    KeyP: 'З',
+    BracketLeft: 'Х',
+    BracketRight: 'Ъ',
+    Backslash: '\\',
+  },
+  ruForCapsFourthRow: {
+    KeyZ: 'Я',
+    KeyX: 'Ч',
+    KeyC: 'С',
+    KeyV: 'М',
+    KeyB: 'И',
+    KeyN: 'Т',
+    KeyM: 'Ь',
+    Comma: 'Б',
+    Period: 'Ю',
+    Slash: '.',
+  },
+};
 
 const container = document.createElement('div');
 container.classList.add('container');
